refactor(mcp): extract evolve_propose tool module
