+++ openspec/changes/refactor-mcp-evolve-propose-tool-module/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-01-22
