+++ openspec/changes/refactor-mcp-evolve-propose-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-evolve-propose-tool-module
+
+Refactor: extract MCP evolve_propose tool implementation into a dedicated module
